Change float to int for consistency when making FTimestamp in FormEntityStatePDU in DISSendComponent

+++ Source/DISRuntime/Private/DISSendComponent.cpp
@@ -202,7 +202,7 @@ FEntityStatePDU UDISSendComponent::FormEntityStatePDU()
 
 	float currentGameTime = GetWorld()->GetTimeSeconds();
 	float totalSeconds;
-	float milliseconds = std::modf(currentGameTime, &totalSeconds) * 1000;
+	int milliseconds = std::modf(currentGameTime, &totalSeconds) * 1000;
 	int minutes = ((int)totalSeconds / 60) % 60;
 	int seconds = (int)totalSeconds % 60;
 	newEntityStatePDU.Timestamp = FTimestamp(ETimestampFormat::Relative, minutes, seconds, milliseconds);
